Merge master (#41)
* Refactor the data set comparison function variable.

There is no need to keep two copies of the data set comparison
function.  This patch adds a method that allows the port code to
obtain the function from the clock code.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>

* telecom: Add the data set comparison algorithm from the Telecom Profiles.

The first half of the telecom algorithm from G.8275.1 and G.8275.2
differs from the one in 1588, making use of profile specific
"localPriority" attributes.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>

* telecom: Introduce the local priority attribute into the data set.

The Telecom Profiles G.8275.1 and G.8275.2 have invented a new
per-port and per-clock attribute, not in 1588, called "localPriority".
The use of this attribute is a distinguishing feature of the telecom
data set comparison algorithm.

This patch adds the attribute, hard coded to its default value.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>

* telecom: Add a configuration option for G.8275.defaultDS.localPriority.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>

* telecom: Add a configuration option for G.8275.portDS.localPriority.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>

* telecom: Add a configuration option to use the alternate BMCA.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>

* telecom: Introduce the masterOnly option.

When masterOnly is true, the port always returns NULL when computing
its best foreign master.  As a result, the port will never enter the
SLAVE state, and the clock will ignore Announce messages received on
that port.

This attribute is specifically called out in G.8275.1 and G.8275.2,
and it is implied by the "master only" mode G.8265.1.  In addition,
this option will probably appear in the next revision of IEEE 1588.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>

* Condense the 802.1AS example configuration.

As we add more and more options, it is getting to be a chore to add
each new option to multiple configuration files.  As we add more
profiles, each with its own example configuration file, the problem
will only get worse.

This patch condenses the gPTP example down to those options which
differ from the built in defaults.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>